ocve/requirements.txt
django-registration-redux
django-tinymce
MySQL-python
xlwt
xlrd
